pyproject.toml
[tool.poetry]
name = "airflow"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11, <3.12"
apache-airflow = "^2.8.1"
pandas = "2.1.0"
apache-airflow-providers-postgres = "^5.10.0"
apache-airflow-providers-amazon = "^8.17.0"
black = "^24.2.0"
astro-sdk-python = {extras = ["amazon", "google", "postgres", "snowflake"], version = "^1.8.0"}
duckdb = "^0.10.0"
dbt-postgres = "^1.7.11"
astronomer-cosmos = "^1.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
